Fix wrong setting of chip status when comic is scrolled out of view
The view holder in the ReadingListItemAdapter would have its status chip set to a different status when it is scrolled out of view. This is due to the fact that ChipGroup's OnCheckedChangeListener is triggered when the Chip's checked state is enabled programmatically. The OnCheckedChangeListener is also used to update the chip's status in the database which is mostly the cause of unwanted status changes when scrolling back to the comic view holder.